Simple example passing
* Home component added to allow index to be wrapper
